+++ load-tests/src/multichain/test_plans/short.txt
@@ -1,0 +1,1 @@
+0,0s;1,120s;0,0s
